[master][537539](UserStory) use autogenerated schema for remote client api (#65)
* [master][537539](UserStory) add autogenerated client api schema

* [master][537539](UserStory) add client api naming overrides

* [master][537539](UserStory) update generated client api code

* [master][537539](UserStory) update sdk for new client api schema

+++ flixpy/flix/extension/extension_api/models/download_request.py
@@ -12,12 +12,12 @@ T = TypeVar("T", bound="DownloadRequest")
 class DownloadRequest:
     """
     Attributes:
-        asset_id (float):
+        asset_id (int):
         target_folder (str):
         asset_type (AssetType):
     """
 
-    asset_id: float
+    asset_id: int
     target_folder: str
     asset_type: AssetType
     additional_properties: Dict[str, Any] = _attrs_field(init=False, factory=dict)
